[acpica] Split AcpiHwLegacySleep.
AcpiHwLegacySleep is called from usermode however part of it must be
executed with interrupts disabled. To handle this we split
AcpiHwLegacySleepFinal into two parts, the part that doesn't need interrupts
disabled, and then the final part (AcpiHwLegacySleepFinal) that does.  We
execute the first half in usermode, and then make a syscall into the kernel to
execute the final steps.

Bug: 76072

Change-Id: I7a5d26a7c11b188b1e133ab66273941ebd6b5cee