[acpica] Jump to the kernel from AcpiHwLegacySleep.
AcpiHwLegacySleep is called from usermode however part of it must be
executed with interrupts disabled. To handle this, when we are running
on Fuchsia and we are not in the kernel, we call on the kernel to
perform the reads and writes of registers that must be done with
interrupts disabled by calling the zx_system_powerctl syscall.

Bug: 76072

Change-Id: I7a5d26a7c11b188b1e133ab66273941ebd6b5cee